% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_epsg.R
\name{ns_epsg}
\alias{ns_epsg}
\title{Set project EPSG value}
\usage{
ns_epsg()
}
\value{
A number matching the EPSG code of the project CRS
}
\description{
Set project EPSG value
}
\examples{
epsg_code <- ns_epsg
}
