% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_exotic.R
\name{ns_exotic}
\alias{ns_exotic}
\title{Exotic status categories}
\usage{
ns_exotic()
}
\value{
A dataframe with two columns and four rows, containing
keys and definitions of exotic status categories acknowledged in floratosLatGr package
}
\description{
Exotic status categories
}
\examples{
ns_exotic()
}
