% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_florichness.R
\name{ns_florichness}
\alias{ns_florichness}
\title{Floristic richness}
\usage{
ns_florichness(ns_df)
}
\arguments{
\item{ns_df}{a dataframe with at least two columns, viz. combination and esotic}
}
\value{
a dataframe with one row and five columns containing number of records,
number of species, number of native, and number of casual, naturalised, and invasive aliens.
}
\description{
Floristic richness
}
\examples{
ns_florichness(floratos_south)
}
