% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_florichness.R
\name{ns_florichness}
\alias{ns_florichness}
\title{Calculate floristi richness from occurrence records}
\usage{
ns_florichness(ns_sf, ns_label)
}
\arguments{
\item{ns_sf}{a dataframe with at least two columns, viz. closest_match and a_status.
Column \code{closest_match} contains scientific names standarised according the
Checklist of the Italian flora. Column \code{a_status} contain the alien status
according to the same Checklist.}

\item{ns_label}{a character string to serve as label for the data in the second column
of the output dataframe.}
}
\value{
a dataframe with six rows and two columns containing number of records,
number of species, number of native, and number of casual, naturalised, and invasive aliens.
}
\description{
Calculate floristi richness from occurrence records
}
\examples{
ns_florichness(floratos_south,"south")
}
